pyproject.toml
[project]
name = "docBasePy"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "nielinjie", email = "nielinjie@gmail.com"},
]
dependencies = [
    "llama-index>=0.10.37",
    "llama-index-embeddings-huggingface>=0.2.0",
    "streamlit>=1.34.0",
    "torch>=2.3.0",
    "transformers>=4.41.0",
    "python-pptx>=0.6.23",
    "Pillow>=10.3.0",
    "streamlit-autorefresh>=1.0.1",
    "python-dotenv>=1.0.1",
]
requires-python = "==3.12.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false
